Make the worker master more generic and resilient to crashes
Currently the worker master is keeping state on all workers in the
system, what they are working on, and for whom they are working. The
weakness of this design is that if the worker master is on a node that
terminates, all of this state is lost. Also the protocol for the worker
master is not generic, and has artifacts related to workflows rather
than any old job message. This update moves the state containing the
requestor of work to the worker and allows workers to register with the
master at any point in the lifecycle - when cluster membership changes,
or during the cadence of completing or failing work.

Also updates the PollerMaster to be abstract as an example